+++ src/main/java/com/ifgoiano/topfilmes/domain/repository/SessionRepository.java
@@ -2,11 +2,12 @@ package com.ifgoiano.topfilmes.domain.repository;
 
 import com.ifgoiano.topfilmes.domain.model.Session;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
 @Repository
 public interface SessionRepository extends JpaRepository<Session, Long> {
-    public Optional<Session> findByCodSession(int codSession);
+    public Optional<Session> findByCodSession(Long codSession);
 }
